+++ frontend/src/components/Chat/Chat.jsx
@@ -4,10 +4,10 @@ import socket from '@/socket';
 import useChatMessage from '@/hooks/useChatMessage';
 
 import Box from '@/components/Common/Box';
-import Form from './Form';
-import MessageList from './MessageList';
+import ChatMessageList from './ChatMessageList';
+import ChatForm from './ChatForm';
 
-export default function Chat({isDonation = true}) {
+export default function Chat({ isDonation = true }) {
     const {
         messageList,
         filterUnsentMessage,
@@ -31,14 +31,14 @@ export default function Chat({isDonation = true}) {
             height="100%"
         >
             <Box width="100%" flex={5} backgroundColor="white" marginBottom={1}>
-                <MessageList
+                <ChatMessageList
                     messageList={messageList}
                     filterUnsentMessage={filterUnsentMessage}
                     deleteMessage={deleteMessage}
                 />
             </Box>
             <Box width="100%" flex={1}>
-                <Form handleSubmit={handleSubmit} isDonation={isDonation} />
+                <ChatForm handleSubmit={handleSubmit} isDonation={isDonation} />
             </Box>
         </Box>
     );
