+++ frontend/src/components/Chat/Chat.stories.jsx
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import Chat from '../Chat';
+import Chat from './Chat';
 
 storiesOf('Chat/Chat', module).add('empty', () => <Chat />);
