+++ frontend/src/components/Chat/ChatMessageList/ChatMessage/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './ChatMessage';
