+++ frontend/src/components/Chat/ChatMessageList/ChatMessageList.stories.jsx
@@ -1,17 +1,16 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import MessageList from '../MessageList';
+import ChatMessageList from './ChatMessageList';
 
 const messageList = [
     { id: 1, type: 'NORMAL', nickname: 'jihoho', content: '쉬워보이네요!' },
     { id: 2, type: 'NORMAL', nickname: 'jihoho', content: '쉬워보이네요!' },
     { id: 3, type: 'NORMAL', nickname: 'jihoho', content: '쉬워보이네요!' },
-    // { type: 'DONATION', nickname: 'jihoho', content: 'BIT_1' },
     { id: 4, type: 'DONATION', nickname: 'jihoho', content: 'BIT_50' },
     { id: 5, type: 'DONATION', nickname: 'jihoho', content: 'BIT_100' },
     { id: 6, type: 'DONATION', nickname: 'jihoho', content: 'BIT_300' },
 ];
 
-storiesOf('Chat/MessageList', module)
-    .add('without props', () => <MessageList />)
-    .add('with props', () => <MessageList messageList={messageList} />);
+storiesOf('Chat/messageList', module)
+    .add('without props', () => <ChatMessageList />)
+    .add('with props', () => <ChatMessageList messageList={messageList} />);
