+++ frontend/src/components/Chat/ChatMessageList/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './ChatMessageList';
